common functions for test
