+++ cdci_data_analysis/pytest_fixtures.py
@@ -13,8 +13,13 @@ import re
 import json
 import string
 import random
+import requests
+import time
+import logging
 
 __this_dir__ = os.path.join(os.path.abspath(os.path.dirname(__file__)))
+
+logger = logging.getLogger()
 
 import signal, psutil
 def kill_child_processes(parent_pid, sig=signal.SIGINT):
@@ -31,6 +36,107 @@ def kill_child_processes(parent_pid, sig=signal.SIGINT):
 def app():
     app = cdci_data_analysis.flask_app.app.app
     return app
+
+
+
+def run_analysis(server, params, method='get'):
+    if method == 'get':
+        return requests.get(server + "/run_analysis",
+                    params={**params},
+                    )
+
+    elif method == 'post':
+        return requests.post(server + "/run_analysis",
+                    data={**params},
+                    )
+    else:
+        raise NotImplementedError
+
+def ask(server, params, expected_query_status, expected_job_status=None, max_time_s=None, expected_status_code=200, method='get'):
+    t0 = time.time()
+
+    c = run_analysis(server, params, method=method)
+
+    logger.info(f"\033[31m request took {time.time() - t0} seconds\033[0m")
+    t_spent = time.time() - t0
+
+    if max_time_s is not None:
+        assert t_spent < max_time_s
+
+    logger.info("content: %s", c.text[:1000])
+    if len(c.text) > 1000:
+        print(".... (truncated)")
+
+    jdata=c.json()
+
+    if expected_status_code is not None:
+        assert c.status_code == expected_status_code
+
+    logger.info(list(jdata.keys()))
+
+    if expected_job_status is not None:
+        assert jdata["exit_status"]["job_status"] in expected_job_status
+
+    if expected_query_status is not None:
+        assert jdata["query_status"] in expected_query_status
+
+    return jdata
+
+def loop_ask(server, params, method='get', max_time_s=None, async_dispatcher=False):
+    jdata = ask(server,
+                {**params, 
+                'async_dispatcher': async_dispatcher,
+                'query_status': 'new',
+                },
+                expected_query_status=["submitted", "done"],
+                method=method,
+                )
+
+    last_status = jdata["query_status"]
+
+    t0 = time.time()
+
+    tries_till_reset = 20
+
+    while True:
+        if tries_till_reset <= 0:
+            next_query_status = "ready"
+            print("\033[1;31;46mresetting query status to new, too long!\033[0m")
+            tries_till_reset = 20
+        else:
+            next_query_status = jdata['query_status']
+            tries_till_reset -= 1
+
+        jdata = ask(server,
+                    {**params, "async_dispatcher": async_dispatcher,
+                            'query_status': next_query_status,
+                            'job_id': jdata['job_monitor']['job_id'],
+                            'session_id': jdata['session_id']},
+                    expected_query_status=["submitted", "done"],
+                    max_time_s=max_time_s,
+                    )
+
+        if jdata["query_status"] in ["ready", "done"]:
+            logger.info("query READY: %s", jdata["query_status"])
+            break
+
+        logger.info("query NOT-READY: %s monitor %s", jdata["query_status"], jdata["job_monitor"])
+        logger.info("looping...")
+
+        time.sleep(5)
+
+    logger.info(f"\033[31m total request took {time.time() - t0} seconds\033[0m")
+
+
+    return jdata, time.time() - t0
+
+def validate_no_data_products(jdata):
+    assert jdata["exit_status"]["debug_message"] == "{\"node\": \"dataanalysis.core.AnalysisException\", \"exception\": \"{}\", \"exception_kind\": \"handled\"}"
+    assert jdata["exit_status"]["error_message"] == "AnalysisException:{}"
+    assert jdata["exit_status"]["message"] == "failed: get dataserver products "
+    assert jdata["job_status"] == "failed"
+
+
 
 
 @pytest.fixture
